Commit works, join up to sync.
Write and commit works. Haven't tested it against multple particpants
because join is not complete, but it should work, let's say it's done.

Snapshots work now. Snapshots will be out of band, with a notification
to the leader to create the snapshot and transport it. Once done, the
leader will invoke a snapshot method which will sync any messages that
where committed while the snapshot took place. Need to sort out the
transition from snapshotting to participating.

I've gone and made this class `async`/`await` and therefore
asynchronous. In the end, it will be deterministic, but it doesn't have
to be synchronous to be deterministic. The only real change from a
synchronous implementation will be logging the notification of the send
loop.

See #5.
Closes #4.