Log new government.
Probably comes in handy to have the full chain of events even we we're
not using the government outside of the half-Paxos, not even in the
ables, but seeing the government go by will be useful. We don't need to
abstract it, I don't think. Perhaps we do not emit the government and
just emit table instructions, like retry after collapse.

See #5.