Bucket actions through paxos.
Buckets now just have a majority and that is available to determine
fallback during departure. Internal state has been reduced. Different
actions put a series of one or two map/reduce actions into paxos. The
actions are performed but the only real state update of the distributor
is to update the majority of the bucket.