Reject state phaser messages.
Removed the old reset logic. Working toward a phaser that can be phased
out of participation and phased back in without an explicit reset
message, just using the paxos method of having an increasing
promise-of-sorts.

Found places where the phaser is combining new and old governments, but
we do all that externally now, so it has been removed.

Now when usurping we assign the government early so that it can be used
to reject messages from the old leader if the old leader is still
around. Otherwise we let the new government go through the two-phase
commit when the leader is expanding.

Closes #20.