Sketch of Sympatico log in JavaScript.
Going to be implementing the Sympatico logic in JavaScript and then
translating that to Rust. Allows me to use a language I know to know
that it is correct and by starting in a green field I'll avoid using
purely JavaScript idioms to express the logic. Will likely use more
classes so I can assemble the Rust implementation from components.