Start of Rust log, use closure instead of push.
Starting the Rust implementation. Already creating generics and working
with closures. Updated JavaScript code to also use closuers.

However, I don't know that this is right pattern. Actually, certain it
is not. Should use a channel or queue of some sort, one that can be used
from another thread and block, so maybe Avenue is the right answer for
the Node.js implementation and we use `VecDeque<T>` for Rust. Otherwise,
at some point I'm going to have to do a lot of type calculus simply to
consume the log, and then probably again to consume the outgoing
messages.